removing components/Image.tsx adding lib/imageslist.ts removing unnecessary css files adding main.css file to add fonts imports changing font size
