+++ app/routes/_index.tsx
@@ -1,12 +1,7 @@
 import { json, type LoaderArgs } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
-import styles from "~/styles/joke.css";
 import { prisma } from "~/lib/prisma.server";
-import Image from "~/components/Image";
-
-export function links() {
-  return [{ rel: "stylesheet", href: styles }];
-}
+import randomImage from "~/lib/imageslist";
 
 export async function loader(params: LoaderArgs) {
   const jokesCount = await prisma.joke.count();
@@ -20,12 +15,17 @@ export async function loader(params: LoaderArgs) {
 
 export default function () {
   const { joke } = useLoaderData();
+  const pic = randomImage();
   return (
-    <div className="w-full bg-gradient-to-b from-indigo-800 to-slate-900 p-4 h-screen flex justify-center items-center">
-      <div className="overflow-hidden absolute top-0 left-0 w-full max-h-full">
-        <Image />
-      </div>
-      <h1 className="joke-content">{joke.content}</h1>
+    <div className="w-full bg-gradient-to-b from-indigo-800 to-slate-900 h-screen flex justify-center items-center">
+      <img
+        alt="laugh"
+        src={pic}
+        className="w-full object-cover h-screen absolute top-0 left-0 z-0"
+      />
+      <h1 className="text-3xl text-center text-white bg-slate-900 rounded-xl z-10 p-4 opacity-80 font-bold max-w-3xl">
+        {joke.content}
+      </h1>
     </div>
   );
 }
