+++ app/styles/main.css
@@ -1,0 +1,7 @@
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Agbalumo&display=swap');
+
+body {
+  margin: 0;
+  font-family: 'Agbalumo', 'Montserrat', sans-serif;
+}
