+++ app/tailwind.css
@@ -1,4 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@100;300&display=swap');
 
 @tailwind base;
 @tailwind components;
